% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eff_frontier}
\alias{eff_frontier}
\title{Efficient Frontier function}
\usage{
eff_frontier(Data)
}
\arguments{
\item{Data}{A portfolio with its parameters to be optimize}
}
\value{
The allocation percent for the portfolio
}
\description{
Takes in any portfolio of stocks return and optimizes it using the Efficient Frontier
}

